restructure all js modules into individual files
