+++ docs/_includes/head.html
@@ -30,7 +30,7 @@
 <script>
     var pageNoPrinterSetting = `{{page.noprinter}}`
 </script>
-<script src="{{ "js/customscripts.js" }}"></script>
+<script type="module" src="{{ "js/customscripts.js" }}"></script>
 
 <link rel="shortcut icon" href="{{ "images/favicon.svg"  }}">
